+++ src/lisa/artifacts/manifest.py
@@ -1,5 +1,6 @@
 import logging
 import os
+from enum import Enum
 from pathlib import Path
 from typing import Annotated, Self
 
@@ -10,11 +11,16 @@ from types_boto3_s3 import S3Client
 logger = logging.getLogger(__name__)
 
 
+class LocalFilesPolicy(str, Enum):
+    OVERWRITE = "overwrite"
+    SKIP = "skip"
+
+
 class Artifact(BaseModel):
     """Represents an artifact with a name and local path."""
 
     name: Annotated[str, Field(min_length=1, strict=True)]
-    """The name of the artifact, usefffd as the S3 key."""
+    """The name of the artifact, used as the S3 key."""
 
     path: Annotated[str, Field(min_length=1, strict=True)]
     """The local file path relative to the local prefix."""
@@ -31,6 +37,11 @@ class ManifestConfiguration(BaseModel):
 
     local_prefix: Annotated[str, Field(strict=True)] = ""
     """Prefix for local paths."""
+
+    local_policy: Annotated[LocalFilesPolicy, Field(strict=True)] = (
+        LocalFilesPolicy.OVERWRITE
+    )
+    """Behaviour when downloading files already present in the local filesystem."""
 
 
 class GetManifestResult(BaseModel):
@@ -69,6 +80,9 @@ class Manifest(BaseModel):
                 bucket=os.environ["ARTIFACTS_BUCKET"],
                 remote_prefix=os.environ.get("ARTIFACTS_REMOTE_PREFIX", ""),
                 local_prefix=os.environ.get("ARTIFACTS_LOCAL_PREFIX", ""),
+                local_policy=LocalFilesPolicy(
+                    os.environ.get("ARTIFACTS_LOCAL_POLICY", LocalFilesPolicy.OVERWRITE)
+                ),
             ),
             artifacts=artifacts,
         )
@@ -112,6 +126,12 @@ class Manifest(BaseModel):
         """
         artifact_key = f"{self.config.remote_prefix}{artifact.name}"
         local_path = Path(self.config.local_prefix) / artifact.path
+        if local_path.is_file():
+            if self.config.local_policy == LocalFilesPolicy.SKIP:
+                logger.info(
+                    "Skipping %s, local file %s exists", artifact_key, local_path
+                )
+                return
         local_path.parent.mkdir(parents=True, exist_ok=True)
         client.download_file(
             Bucket=self.config.bucket,