+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added local files policy to manifest configuration.
+
 ## [0.3.0] - 2026-01-14
 
 ### Added